Fixes some warnings when compiling.

+++ test/mfem/Forms.cpp
@@ -135,7 +135,7 @@ TEST_F(FormTest, MassMatrixToSingleBlockMatrix)
 		break;
 	case (Element::Type::QUADRILATERAL):
 		ne = nx * ny;
-		dof_per_order = std::pow(order + 1, 2);
+		dof_per_order = int(std::pow(order + 1, 2));
 		break;
 	}
